Updated crt-lottes to include his latest changes
changes include the addition of halation/bloom and several options for shadow mask / aperture grille. I put the bloom into the runtime parameters, but I couldn't think of a good way to do the masks without if-branching all of it, which would kill performance. Maybe aliaspider has some input on that.